MAME-ification Update.
See changelog.

+++ aae/aae/vidhrdwr/rallyx_vid.cpp
@@ -48,7 +48,7 @@ void rallyx_vh_convert_color_prom(unsigned char* palette, unsigned char* colorta
 #define TOTAL_COLORS(gfxn) (Machine->gfx[gfxn]->total_colors * Machine->gfx[gfxn]->color_granularity)
 #define COLOR(gfxn,offs) (colortable[Machine->drv->gfxdecodeinfo[gfxn].color_codes_start + offs])
 
-	for (i = 0; i < Machine->drv->total_colors; i++)
+	for (i = 0; i < (int) Machine->drv->total_colors; i++)
 	{
 		int bit0, bit1, bit2;
